fix(UI): chart labels clip and axis range fix
Change text_position to auto for Plotly to decide the
best layout for chart labels. Chart labels are now
located inside the bars.

Order cancellation and fulfillment friction section's
bar graph's vertical axis has a maximum value of 10%
by default. Before this, the bars that were below 5%
(which is most of the bars) were not clearly visible
as they were too small.